Enable using texify for inline math
